requirements.txt
aio-pika==9.0.4
aiohttp==3.8.3
aiohttp-apispec==2.2.3
aiohttp-session==2.12.0
aiormq==6.7.2
aiosignal==1.3.1
alembic==1.9.3
apispec==3.3.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
black==23.1.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==7.1.0
cryptography==39.0.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pamqp==3.2.1
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pycparser==2.21
pytest==7.2.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
webargs==5.5.3
yarl==1.8.2
